+++ tests/Monolog/HandlerTest.php
@@ -20,28 +20,32 @@ final class HandlerTest extends TestCase
     /**
      * @dataProvider handleDataProvider
      */
-    public function testHandle(array $record, Event $expectedEvent, EventHint $expectedHint, array $expectedExtra): void
+    public function testHandle(bool $fillExtraContext, array $record, Event $expectedEvent, EventHint $expectedHint, array $expectedExtra): void
     {
         /** @var ClientInterface&MockObject $client */
         $client = $this->createMock(ClientInterface::class);
         $client->expects($this->once())
             ->method('captureEvent')
-            ->with($this->callback(function (Event $event) use ($expectedEvent): bool {
-                $this->assertEquals($expectedEvent->getLevel(), $event->getLevel());
-                $this->assertEquals($expectedEvent->getMessage(), $event->getMessage());
-                $this->assertEquals($expectedEvent->getLogger(), $event->getLogger());
+            ->with(
+                $this->callback(function (Event $event) use ($expectedEvent): bool {
+                    $this->assertEquals($expectedEvent->getLevel(), $event->getLevel());
+                    $this->assertSame($expectedEvent->getMessage(), $event->getMessage());
+                    $this->assertSame($expectedEvent->getLogger(), $event->getLogger());
 
-                return true;
-            }), $expectedHint, $this->callback(function (Scope $scopeArg) use ($expectedExtra): bool {
-                $event = $scopeArg->applyToEvent(Event::createEvent());
+                    return true;
+                }),
+                $expectedHint,
+                $this->callback(function (Scope $scopeArg) use ($expectedExtra): bool {
+                    $event = $scopeArg->applyToEvent(Event::createEvent());
 
-                $this->assertNotNull($event);
-                $this->assertSame($expectedExtra, $event->getExtra());
+                    $this->assertNotNull($event);
+                    $this->assertSame($expectedExtra, $event->getExtra());
 
-                return true;
-            }));
+                    return true;
+                })
+            );
 
-        $handler = new Handler(new Hub($client, new Scope()));
+        $handler = new Handler(new Hub($client, new Scope()), Logger::DEBUG, true, $fillExtraContext);
         $handler->handle($record);
     }
 
@@ -53,6 +57,7 @@ final class HandlerTest extends TestCase
         $event->setLevel(Severity::debug());
 
         yield [
+            false,
             [
                 'message' => 'foo bar',
                 'level' => Logger::DEBUG,
@@ -75,6 +80,7 @@ final class HandlerTest extends TestCase
         $event->setLevel(Severity::info());
 
         yield [
+            false,
             [
                 'message' => 'foo bar',
                 'level' => Logger::INFO,
@@ -97,6 +103,7 @@ final class HandlerTest extends TestCase
         $event->setLevel(Severity::info());
 
         yield [
+            false,
             [
                 'message' => 'foo bar',
                 'level' => Logger::NOTICE,
@@ -119,6 +126,7 @@ final class HandlerTest extends TestCase
         $event->setLevel(Severity::warning());
 
         yield [
+            false,
             [
                 'message' => 'foo bar',
                 'level' => Logger::WARNING,
@@ -141,6 +149,7 @@ final class HandlerTest extends TestCase
         $event->setLevel(Severity::error());
 
         yield [
+            false,
             [
                 'message' => 'foo bar',
                 'level' => Logger::ERROR,
@@ -163,6 +172,7 @@ final class HandlerTest extends TestCase
         $event->setLevel(Severity::fatal());
 
         yield [
+            false,
             [
                 'message' => 'foo bar',
                 'level' => Logger::CRITICAL,
@@ -185,6 +195,7 @@ final class HandlerTest extends TestCase
         $event->setLevel(Severity::fatal());
 
         yield [
+            false,
             [
                 'message' => 'foo bar',
                 'level' => Logger::ALERT,
@@ -207,6 +218,7 @@ final class HandlerTest extends TestCase
         $event->setLevel(Severity::fatal());
 
         yield [
+            false,
             [
                 'message' => 'foo bar',
                 'level' => Logger::EMERGENCY,
@@ -231,6 +243,7 @@ final class HandlerTest extends TestCase
         $exampleException = new \Exception('exception message');
 
         yield [
+            false,
             [
                 'message' => 'foo bar',
                 'level' => Logger::WARNING,
@@ -245,6 +258,89 @@ final class HandlerTest extends TestCase
             EventHint::fromArray([
                 'exception' => $exampleException,
             ]),
+            [
+                'monolog.channel' => 'channel.foo',
+                'monolog.level' => Logger::getLevelName(Logger::WARNING),
+            ],
+        ];
+
+        $event = Event::createEvent();
+        $event->setMessage('foo bar');
+        $event->setLogger('monolog.channel.foo');
+        $event->setLevel(Severity::warning());
+
+        yield 'Monolog\'s context is filled and the handler should fill the "extra" context' => [
+            true,
+            [
+                'message' => 'foo bar',
+                'level' => Logger::WARNING,
+                'level_name' => Logger::getLevelName(Logger::WARNING),
+                'context' => [
+                    'foo' => 'bar',
+                    'bar' => 'baz',
+                ],
+                'channel' => 'channel.foo',
+                'extra' => [],
+            ],
+            $event,
+            new EventHint(),
+            [
+                'monolog.channel' => 'channel.foo',
+                'monolog.level' => Logger::getLevelName(Logger::WARNING),
+                'monolog.context' => [
+                    'foo' => 'bar',
+                    'bar' => 'baz',
+                ],
+            ],
+        ];
+
+        $event = Event::createEvent();
+        $event->setMessage('foo bar');
+        $event->setLogger('monolog.channel.foo');
+        $event->setLevel(Severity::warning());
+
+        yield 'Monolog\'s context is filled with "exception" field and the handler should fill the "extra" context' => [
+            true,
+            [
+                'message' => 'foo bar',
+                'level' => Logger::WARNING,
+                'level_name' => Logger::getLevelName(Logger::WARNING),
+                'context' => [
+                    'exception' => new \Exception('exception message'),
+                ],
+                'channel' => 'channel.foo',
+                'extra' => [],
+            ],
+            $event,
+            EventHint::fromArray([
+                'exception' => $exampleException,
+            ]),
+            [
+                'monolog.channel' => 'channel.foo',
+                'monolog.level' => Logger::getLevelName(Logger::WARNING),
+            ],
+        ];
+
+        $event = Event::createEvent();
+        $event->setMessage('foo bar');
+        $event->setLogger('monolog.channel.foo');
+        $event->setLevel(Severity::warning());
+
+        yield 'Monolog\'s context is filled but handler should not fill the "extra" context' => [
+            false,
+            [
+                'message' => 'foo bar',
+                'level' => Logger::WARNING,
+                'level_name' => Logger::getLevelName(Logger::WARNING),
+                'context' => [
+                    'foo' => 'bar',
+                    'bar' => 'baz',
+                ],
+                'channel' => 'channel.foo',
+                'extra' => [],
+            ],
+            $event,
+            new EventHint(),
             [
                 'monolog.channel' => 'channel.foo',
                 'monolog.level' => Logger::getLevelName(Logger::WARNING),